+++ opencti-platform/opencti-front/src/private/components/settings/sso_definitions/SSODefinitionForm.tsx
@@ -17,7 +17,8 @@ import SAMLConfig from '@components/settings/sso_definitions/SAMLConfig';
 import OpenIDConfig from '@components/settings/sso_definitions/OpenIDConfig';
 import { ConfigurationTypeInput } from '@components/settings/sso_definitions/__generated__/SSODefinitionCreationMutation.graphql';
 import Button from '@common/button/Button';
-import IconButton from '@common/button/IconButton';
+import SSODefinitionGroupForm from '@components/settings/sso_definitions/SSODefinitionGroupForm';
+import SSODefinitionOrganizationForm from '@components/settings/sso_definitions/SSODefinitionOrganizationForm';
 
 interface SSODefinitionFormProps {
   onCancel: () => void;
@@ -207,15 +208,6 @@ const SSODefinitionForm = ({
     if (onSubmitField) onSubmitField(field, value);
   };
 
-  const getGroupAttributeKeyName = () => {
-    switch (selectedStrategy) {
-      case 'SAML': return 'group_attributes';
-      case 'OpenID': return 'groups_attributes';
-      case 'LDAP': return 'group_attribute';
-      default: return '';
-    }
-  };
-
   return (
     <Formik
       enableReinitialize={!updateField}
@@ -281,204 +273,8 @@ const SSODefinitionForm = ({
               {selectedStrategy === 'OpenID' && <OpenIDConfig updateField={updateField} />}
             </>
           )}
-          {currentTab === 1 && (
-            <>
-              <Field
-                sx={{ marginTop: '20px' }}
-                component={TextField}
-                variant="standard"
-                name={getGroupAttributeKeyName()}
-                onSubmit={updateField}
-                label={t_i18n('Attribute in token')}
-                containerstyle={{ marginTop: 12 }}
-                fullWidth
-              />
-              {selectedStrategy === 'OpenID' && (
-                <Field
-                  sx={{ marginTop: '20px' }}
-                  component={TextField}
-                  variant="standard"
-                  name="groups_path"
-                  onSubmit={updateField}
-                  label={t_i18n('Group path')}
-                  containerstyle={{ marginTop: 12 }}
-                  fullWidth
-                />
-              )}
-              <FieldArray name="groups_mapping">
-                {({ push, remove, form }) => (
-                  <>
-                    <div style={{ display: 'flex', alignItems: 'center', marginTop: 20 }}>
-                      <Typography variant="h2">{t_i18n('Add a new value')}</Typography>
-                      <IconButton
-                        size="default"
-                        color="secondary"
-                        aria-label={t_i18n('Add a new value')}
-                        style={{ marginBottom: 12 }}
-                        onClick={() => push('')}
-                      >
-                        <Add fontSize="small" color="primary" />
-                      </IconButton>
-                    </div>
-                    {form.values.groups_mapping
-                      && form.values.groups_mapping.map(
-                        (value: string, index: number) => (
-                          <div
-                            key={index}
-                            style={{
-                              display: 'flex',
-                              alignItems: 'center',
-                              justifyContent: 'space-between',
-                              marginBottom: 8,
-                            }}
-                          >
-                            <Field
-                              component={TextField}
-                              variant="standard"
-                              onSubmit={() => updateField('groups_mapping', form.values.groups_mapping)}
-                              name={`groups_mapping[${index}]`}
-                              label={t_i18n('Group mapping value')}
-                              fullWidth
-                              style={{ marginTop: 20 }}
-                            />
-                            {/* <div */}
-                            {/*  style={{ */}
-                            {/*    flexBasis: '70%', */}
-                            {/*    maxWidth: '70%', */}
-                            {/*    marginBottom: 20, */}
-                            {/*  }} */}
-                            {/* > */}
-                            {/*  <GroupField */}
-                            {/*    name="groups" */}
-                            {/*    label="Groups" */}
-                            {/*    style={fieldSpacingContainerStyle} */}
-                            {/*    showConfidence={true} */}
-                            {/*  /> */}
-                            {/* </div> */}
-                            <IconButton
-                              size="default"
-                              color="primary"
-                              aria-label={t_i18n('Delete')}
-                              style={{ marginTop: 10 }}
-                              onClick={() => {
-                                const groupsMapping = [...form.values.groups_mapping];
-                                groupsMapping.splice(index, 1);
-                                remove(index);
-                                updateField('groups_mapping', groupsMapping);
-                              }} // Delete
-                            >
-                              <Delete fontSize="small" />
-                            </IconButton>
-                            {/* <Field */}
-                            {/*  component={SwitchField} */}
-                            {/*  variant="standard" */}
-                            {/*  type="checkbox" */}
-                            {/*  name="auto_create_group" */}
-                            {/*  label={t_i18n('auto-create group')} */}
-                            {/*  containerstyle={{ marginTop: 10 }} */}
-                            {/* /> */}
-                          </div>
-                        ),
-                      )}
-                  </>
-                )}
-              </FieldArray>
-              {/* <Field */}
-              {/*  component={SwitchField} */}
-              {/*  variant="standard" */}
-              {/*  type="checkbox" */}
-              {/*  name="read_userinfo" */}
-              {/*  label={t_i18n('Automatically add users to default groups')} */}
-              {/*  containerstyle={{ marginLeft: 2, marginTop: 30 }} */}
-              {/* /> */}
-            </>
-          )}
-          {currentTab === 2 && (
-            <>
-              <div style={{ marginTop: 20 }}>
-                <Field
-                  component={TextField}
-                  variant="standard"
-                  name="organizations_path"
-                  onSubmit={updateField}
-                  label={t_i18n('Attribute/path in token')}
-                  fullWidth
-                />
-              </div>
-              <FieldArray name="organizations_mapping">
-                {({ push, remove, form }) => (
-                  <>
-                    <div
-                      style={{
-                        display: 'flex',
-                        alignItems: 'center',
-                        marginTop: 20,
-                      }}
-                    >
-                      <Typography variant="h2">{t_i18n('Add a new value')}</Typography>
-                      <IconButton
-                        size="default"
-                        color="secondary"
-                        aria-label={t_i18n('Add a new value')}
-                        style={{ marginBottom: 12 }}
-                        onClick={() => push('')}
-                      >
-                        <Add fontSize="small" color="primary" />
-                      </IconButton>
-                    </div>
-                    {form.values.organizations_mapping
-                      && form.values.organizations_mapping.map(
-                        (value: string, index: number) => (
-                          <div
-                            key={index}
-                            style={{
-                              display: 'flex',
-                              alignItems: 'flex-start',
-                              marginBottom: 8,
-                            }}
-                          >
-                            <Field
-                              component={TextField}
-                              variant="standard"
-                              name={`organizations_mapping[${index}]`}
-                              label={t_i18n('Value organizations mappings')}
-                              onSubmit={() => updateField('organizations_mapping', form.values.organizations_mapping)}
-                              fullWidth
-                              style={{ marginTop: 20 }}
-                            />
-                            {/* <div */}
-                            {/*  style={{ flexBasis: '70%', maxWidth: '70%' }} */}
-                            {/* > */}
-                            {/*  <ObjectOrganizationField */}
-                            {/*    outlined={false} */}
-                            {/*    name="objectOrganization" */}
-                            {/*    label="Organizations" */}
-                            {/*    containerstyle={{ width: '100%' }} */}
-                            {/*    style={fieldSpacingContainerStyle} */}
-                            {/*    fullWidth */}
-                            {/*  /> */}
-                            {/* </div> */}
-                            <IconButton
-                              color="primary"
-                              aria-label={t_i18n('Delete')}
-                              style={{ marginTop: 30, marginLeft: 50 }}
-                              onClick={() => {
-                                remove(index);
-                                const organizationsMapping = [...form.values.organizations_mapping];
-                                organizationsMapping.splice(index, 1);
-                                updateField('organizations_mapping', organizationsMapping);
-                              }}
-                            >
-                              <Delete fontSize="small" />
-                            </IconButton>
-                          </div>
-                        ),
-                      )}
-                  </>
-                )}
-              </FieldArray>
-            </>
-          )}
+          {currentTab === 1 && <SSODefinitionGroupForm updateField={updateField} selectedStrategy={selectedStrategy} />}
+          {currentTab === 2 && <SSODefinitionOrganizationForm updateField={updateField} />}
           {!onSubmitField && (
             <div
               style={{
